fix: wrong type for price before float after bigDecimal

+++ api_inventory/src/main/java/com/inventory_service/api/dto/ResponseProductDTO.java
@@ -2,6 +2,7 @@ package com.inventory_service.api.dto;
 
 import com.inventory_service.api.model.ProductModel;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -9,7 +10,7 @@ public record ResponseProductDTO(
         UUID id,
         String name,
         String sku,
-        Float price,
+        BigDecimal price,
         int stock,
         LocalDateTime created_at
 ) {
